my_project/app/requirements.txt
Flask
psycopg2-binary
redis
elasticsearch